[Feat] support anthropics evals dataset
