+++ configs/datasets/anthropics_evals/persona_gen.py
@@ -1,0 +1,4 @@
+from mmengine.config import read_base
+
+with read_base():
+    from .persona_gen_cc72e2 import persona_datasets  # noqa: F401, F403
